Update build (#74)
* Set the c++ standard in the build files
* Install all conda dependencies at once
  + Update dependency minor versions
  + Add commented out dependencies for Python 2
* Check environment variables for compilers in build files
* Address issues previously patched in the bioconda recipe
  https://github.com/bioconda/bioconda-recipes/tree/4d8723c5b52dd4310934da589de887198784e7b6/recipes/rmats
* Update README
* Address warning messages from build output

+++ bamtools/src/api/internal/io/BamHttp_p.cpp
@@ -406,7 +406,7 @@ bool BamHttp::SendGetRequest(const size_t numBytes) {
     // send request
     const string requestHeader = m_request->ToString();
     const size_t headerSize    = requestHeader.size();
-    if ( WriteToSocket(requestHeader.c_str(), headerSize) != headerSize ) {
+    if ( WriteToSocket(requestHeader.c_str(), headerSize) != (int64_t)headerSize ) {
         SetErrorString("BamHttp::SendHeadRequest", m_socket->GetErrorString());
         return false;
     }
@@ -498,7 +498,7 @@ bool BamHttp::SendHeadRequest(void) {
     // send request
     const string requestHeader = m_request->ToString();
     const size_t headerSize    = requestHeader.size();
-    if ( WriteToSocket(requestHeader.c_str(), headerSize) != headerSize ) {
+    if ( WriteToSocket(requestHeader.c_str(), headerSize) != (int64_t)headerSize ) {
         SetErrorString("BamHttp::SendHeadRequest", m_socket->GetErrorString());
         return false;
     }
